JD- Card Hover Logic
Reworked Card Hover Logic,
it sort of works, just needs to be changed to mimic the card. I can explain this further. Need to work with Chelsea to perfect it.

We have the option of a temp fix for the hack by just spreading the hand out further but this is a decent jumping off point.

I've left the old code commented out JIC we need it..