Enemy, Player, Turn Logic
Started implementing the joint interactions between enemies, the player and the game itself.

Added a turnCounter as a go between and methods to EnemyLogic to allow it to be told to start its turn. The enemies themselves will end their turn, this WILL CAUSE AN ERROR CURRENTLY BECAUSE EACH ENEMY WILL RUN ITS TURN AND INDEPENDENTLY END ALL ENEMIES TURNS!! (JD to fix)