Pictures updated for the Tutorial page and bug fixed  enemy defence
Pictures updated for the Tutorial page and bug fixed for enemy defence ui. As before when the enemy plays a defence card. The UI text will change from numbers to formate and disapper. To fix this i updated the code on the enemy script, as there was a secrtion in the defence method that needed the new code.